kernel/.../{kprove,krun,unparser}: separate out KPrint module for encapsulating output information
