+++ kernel/src/main/java/org/kframework/krun/KRun.java
@@ -1,45 +1,26 @@
 // Copyright (c) 2015-2018 K Team. All Rights Reserved.
 package org.kframework.krun;
 
-import com.davekoelle.AlphanumComparator;
 import org.apache.commons.lang3.tuple.Pair;
-import org.kframework.attributes.Att;
 import org.kframework.attributes.Source;
 import org.kframework.builtin.KLabels;
 import org.kframework.builtin.Sorts;
 import org.kframework.compile.ConfigurationInfoFromModule;
-import org.kframework.compile.ExpandMacros;
-import org.kframework.definition.Definition;
 import org.kframework.definition.Module;
 import org.kframework.definition.Rule;
 import org.kframework.kompile.CompiledDefinition;
-import org.kframework.kompile.KompileOptions;
 import org.kframework.kore.Assoc;
 import org.kframework.kore.K;
 import org.kframework.kore.KApply;
-import org.kframework.kore.KORE;
 import org.kframework.kore.KToken;
 import org.kframework.kore.KVariable;
 import org.kframework.kore.Sort;
-import org.kframework.kore.TransformK;
-import org.kframework.kore.VisitK;
 import org.kframework.krun.modes.ExecutionMode;
-import org.kframework.main.GlobalOptions;
 import org.kframework.main.Main;
-import org.kframework.parser.ProductionReference;
 import org.kframework.parser.binary.BinaryParser;
-import org.kframework.parser.concrete2kore.ParseInModule;
-import org.kframework.parser.concrete2kore.generator.RuleGrammarGenerator;
 import org.kframework.parser.kore.KoreParser;
 import org.kframework.rewriter.Rewriter;
-import org.kframework.unparser.AddBrackets;
-import org.kframework.unparser.ColorSetting;
-import org.kframework.unparser.Formatter;
-import org.kframework.unparser.KOREToTreeNodes;
-import org.kframework.unparser.OutputModes;
-import org.kframework.unparser.PrintOptions;
-import org.kframework.unparser.ToBinary;
-import org.kframework.unparser.ToKast;
+import org.kframework.unparser.KPrint;
 import org.kframework.utils.errorsystem.KEMException;
 import org.kframework.utils.errorsystem.KException;
 import org.kframework.utils.errorsystem.KExceptionManager;
@@ -53,15 +34,12 @@ import java.io.IOException;
 import java.io.InputStreamReader;
 import java.util.ArrayList;
 import java.util.Arrays;
-import java.util.Comparator;
 import java.util.HashMap;
 import java.util.HashSet;
 import java.util.List;
 import java.util.Map;
 import java.util.Set;
-import java.util.function.Consumer;
 import java.util.function.Function;
-import java.util.stream.Collectors;
 
 import static org.kframework.Collections.*;
 import static org.kframework.kore.KORE.*;
@@ -108,7 +86,7 @@ public class KRun {
 
 
         if (result != null) {
-            prettyPrint(compiledDef.getParsedDefinition(), compiledDef.languageParsingModule(), files, compiledDef.kompileOptions, options.print.output, s -> outputFile(s, options.print, files), result._1(), options.print.color(tty.stdout, files.getEnv()));
+            KPrint.prettyPrint(compiledDef.getParsedDefinition(), compiledDef.languageParsingModule(), files, compiledDef.kompileOptions, options.print.output, s -> KPrint.outputFile(s, options.print, files), result._1(), options.print.color(tty.stdout, files.getEnv()));
             return result._2();
         }
         return 0;
@@ -146,23 +124,6 @@ public class KRun {
         return vars.iterator().next();
     }
 
-    //TODO(dwightguth): use Writer
-    public static void outputFile(String output, PrintOptions options, FileUtil files) {
-        outputFile(output.getBytes(), options, files);
-    }
-
-    public static void outputFile(byte[] output, PrintOptions options, FileUtil files) {
-        if (options.outputFile == null) {
-            try {
-                System.out.write(output);
-            } catch (IOException e) {
-                throw KEMException.internalError(e.getMessage(), e);
-            }
-        } else {
-            files.saveToWorkingDirectory(options.outputFile, output);
-        }
-    }
-
     /**
      * Function to compile the String Pattern, if the pattern is not present in the cache. Note the difference between
      * compilation and parsing. Compilation is the result of resolving anonymous variables, semantic casts, and concretizing
@@ -186,31 +147,6 @@ public class KRun {
      */
     public static Rule parsePattern(FileUtil files, KExceptionManager kem, String pattern, CompiledDefinition compiledDef, Source source) {
         return compiledDef.parsePatternIfAbsent(files, kem, pattern, source);
-    }
-
-    public static void prettyPrint(Definition def, Module module, FileUtil files, KompileOptions kompile, OutputModes output, Consumer<byte[]> print, K result, ColorSetting colorize) {
-        switch (output) {
-        case KAST:
-            print.accept((ToKast.apply(result) + "\n").getBytes());
-            break;
-        case NONE:
-            print.accept("".getBytes());
-            break;
-        case PRETTY: {
-            Module unparsingModule = RuleGrammarGenerator.getCombinedGrammar(module, false).getExtensionModule();
-            print.accept((unparseTerm(result, unparsingModule, colorize, files, kompile) + "\n").getBytes());
-            break;
-        } case PROGRAM: {
-            RuleGrammarGenerator gen = new RuleGrammarGenerator(def);
-            Module unparsingModule = RuleGrammarGenerator.getCombinedGrammar(gen.getProgramsGrammar(module), false).getParsingModule();
-            print.accept((unparseTerm(result, unparsingModule, colorize, files, kompile) + "\n").getBytes());
-            break;
-        } case BINARY:
-            print.accept(ToBinary.apply(result));
-            break;
-        default:
-            throw KEMException.criticalError("Unsupported output mode: " + output);
-        }
     }
 
     private K parseConfigVars(KRunOptions options, CompiledDefinition compiledDef) {
@@ -282,49 +218,6 @@ public class KRun {
 
     public KApply plugConfigVars(CompiledDefinition compiledDef, Map<KToken, K> output) {
         return KApply(compiledDef.topCellInitializer, output.entrySet().stream().map(e -> KApply(KLabel("_|->_"), e.getKey(), e.getValue())).reduce(KApply(KLabel(".Map")), (a, b) -> KApply(KLabel("_Map_"), a, b)));
-    }
-
-    public static String unparseTerm(K input, Module test, ColorSetting colorize, FileUtil files, KompileOptions kompile) {
-        K sortedComm = new TransformK() {
-            @Override
-            public K apply(KApply k) {
-                if (k.klabel() instanceof KVariable) {
-                    return super.apply(k);
-                }
-                Att att = test.attributesFor().apply(KLabel(k.klabel().name()));
-                if (att.contains("comm") && att.contains("assoc") && att.contains("unit")) {
-                    List<K> items = new ArrayList<>(Assoc.flatten(k.klabel(), k.klist().items(), KLabel(att.get("unit"))));
-                    List<Tuple2<String, K>> printed = new ArrayList<>();
-                    for (K item : items) {
-                        String s = unparseInternal(test, ColorSetting.OFF, apply(item), files, kompile);
-                        printed.add(Tuple2.apply(s, item));
-                    }
-                    printed.sort(Comparator.comparing(Tuple2::_1, new AlphanumComparator()));
-                    items = printed.stream().map(Tuple2::_2).map(this::apply).collect(Collectors.toList());
-                    return items.stream().reduce((k1, k2) -> KApply(k.klabel(), k1, k2)).orElse(KApply(KLabel(att.get("unit"))));
-                }
-                return super.apply(k);
-            }
-        }.apply(input);
-        K alphaRenamed = new TransformK() {
-            Map<KVariable, KVariable> renames = new HashMap<>();
-            int newCount = 0;
-
-            @Override
-            public K apply(KVariable k) {
-                if (k.att().contains("anonymous")) {
-                    return renames.computeIfAbsent(k, k2 -> KVariable("V" + newCount++, k.att()));
-                }
-                return k;
-            }
-        }.apply(sortedComm);
-        return unparseInternal(test, colorize, alphaRenamed, files, kompile);
-    }
-
-    private static String unparseInternal(Module mod, ColorSetting colorize, K input, FileUtil files, KompileOptions kompile) {
-        ExpandMacros expandMacros = new ExpandMacros(mod, files, kompile, true);
-        return Formatter.format(
-                new AddBrackets(mod).addBrackets((ProductionReference) ParseInModule.disambiguateForUnparse(mod, KOREToTreeNodes.apply(KOREToTreeNodes.up(mod, expandMacros.expand(input)), mod))), colorize);
     }
 
     public K externalParse(String parser, String value, Sort startSymbol, Source source, CompiledDefinition compiledDef) {
